Start adding a new analsysis that collects tags on functions
The idea is to allow function definitions and declarations to have a
__tags("foo","bar",...) attribute which we collect and then have available for
later passes.

for example we could have

```c
```

and later we could have

```c
/* in a header */
MONO_API void mono_do_something (void);
```

```c
/* in the implementation */
void
mono_do_something (void)
{
  /* whatever */
}
```

In later passes we could query "mono_do_something" and learn that it has the
tag "mono_api".

The implementation is flexible so that if there are multiple declarations (or
if the definition is also tagged), all the tags are unioned together.